fix(bench): drop useless static_cast on state.iterations()
Google Benchmark's IterationCount typedef alias is std::int64_t on libstdc++; wrapping state.iterations() in static_cast<std::int64_t> trips -Wuseless-cast under GCC. Keep the cast on the std::size_t multiplier (still needed for -Wsign-conversion) and drop it on the iteration count.

+++ bench/bench_arena.cpp
@@ -33,7 +33,7 @@ void bench_alloc_only(benchmark::State& state) {
         live.clear();
         state.ResumeTiming();
     }
-    state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+    state.SetItemsProcessed(state.iterations() *
                             static_cast<std::int64_t>(n));
 }
 
@@ -54,7 +54,7 @@ void bench_dealloc_only(benchmark::State& state) {
         live.clear();
         benchmark::ClobberMemory();
     }
-    state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+    state.SetItemsProcessed(state.iterations() *
                             static_cast<std::int64_t>(n));
 }
